==> 0.1 Meta-analysis_Acute.Rmd <==
---
title: "Meta-analysis"
output: html_document
date: "2025-02-26"
---
 
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F)
```

Packages used: 

```{r}
library(Seurat)
library(Matrix)
library(biomaRt)
library(AnnotationDbi)
library(org.Mm.eg.db)
library(ggplot2)
library(dplyr)
library(stringr)
library(readxl)
library(openxlsx)
library(patchwork)
library(tibble)
```

1. First of all, we need to prepare the three datasets for integration. One of the most important points is standarize the labels of the dataset.The different datasets are already filtered by QC metrics. (% mitochondrial < 0.2, no doublets)

Dataset Milich 

```{r}
matrix <- readMM("C:/Users/afornies/Downloads/Dataset1.mtx")
metadata <- read.delim("C:/Users/afornies/Downloads/metadata1.tsv", header = T)
features <- read.delim("C:/Users/afornies/Downloads/features1.tsv")
barcodes <- rownames(metadata)
```

```{r}
metadata <- metadata %>% dplyr::select(nCount_RNA, nFeature_RNA, time, celltype, sample_id)

metadata$technology <- "scRNA"
metadata$dataset <- "Milich_2021"

rownames(matrix) <- features$mgi_symbol
colnames(matrix) <- rownames(barcodes)
```

```{r}
seurat <- CreateSeuratObject(matrix,assay = "RNA", project = "Milich_2021", meta.data = metadata, min.cells = 10, min.features = 100)

remove_features <- nrow(matrix) - nrow(seurat)
remove_cells <- ncol(matrix) - ncol(seurat)

seurat <- subset(seurat, subset = time %in% c("Uninjured", "1dpi", "7dpi"))

remove_cells_f <- ncol(matrix) - nrow(seurat)

rm(matrix,barcodes,metadata,features)
```

Dataset Matson 2022

```{r}
matrix2 <- readRDS("C:/Users/afornies/Downloads/Dataset2.rds")
```

```{r}
counts2 <- matrix2@assays[["RNA"]]@counts

metadata2 <- matrix2@meta.data

metadata2$orig.ident <- "Matson_2022 (Single cell Atlas)"

colnames(metadata2)[colnames(metadata2) == "coarse_clusters"] <- "celltype"
colnames(metadata2)[colnames(metadata2) == "condition"] <- "time"
colnames(metadata2)[colnames(metadata2) == "nGene"] <- "nFeature_RNA"
colnames(metadata2)[colnames(metadata2) == "nUMI"] <- "nCount_RNA"

metadata2$time <- gsub("C_1wpi", "7dpi", metadata2$time)
metadata2$time <- gsub("A_Uninj", "Uninjured", metadata2$time)
metadata2$time <- gsub("B_1dpi", "1dpi", metadata2$time)

colnames(metadata2) <- make.unique(colnames(metadata2))

metadata2 <- metadata2 %>% select(matches("nCount_RNA|nFeature_RNA|time|celltype"))

metadata2 <- metadata2 [, -c(3,4)]

metadata2$technology <- "snRNA"
metadata2$dataset <- "Matson_2022"

```

```{r}
seurat2 <- CreateSeuratObject(counts2, assay = "RNA", project = "Matson_2022 (Single cell Atlas)", meta.data = metadata2, min.cells = 10, min.features = 100)

remove_features2 <- nrow(matrix2) - nrow(seurat2)
remove_cells2 <- ncol(matrix2) - ncol(seurat2)

seurat2 <- subset(seurat2, subset = time %in% c("Uninjured", "1dpi", "7dpi"))

remove_cells2_f <- ncol(matrix2) - ncol(seurat2)

rm(matrix2, counts2, metadata2)
```

Dataset Skinnider_2024

```{r}

matrix3 <- readMM("C:/Users/afornies/Downloads/Dataset3.mtx")
gc()
barcodes3 <- read.delim("C:/Users/afornies/Downloads/barcodes3.txt.gz", header = F)
metadata3 <- read.delim("C:/Users/afornies/Downloads/metadata3.txt", header = T)
features3 <- read.delim("C:/Users/afornies/Downloads/features3.txt.gz", header = F)

```


```{r}

colnames(matrix3) <- barcodes3$V1
rownames(matrix3) <- features3$V1

rownames(metadata3) <- metadata3$barcode

matrix3 <- as(matrix3, "dgCMatrix")
gc()

metadata3 <- metadata3 %>% rename(time = label)

metadata3$time <- recode(metadata3$time, 
                         "contusion" = "7dpi", 
                         "uninjured" = "Uninjured")

metadata3 <- metadata3 %>% rename(celltype = layer2)

metadata3 <- metadata3 %>% select(matches("nCount_RNA|nFeature_RNA|time|celltype"))

metadata3$technology <- "snRNA"
metadata3$dataset <- "Skinnider_2024"

```

```{r}
seurat3  <- CreateSeuratObject(matrix3, assay = "RNA", project = "Skinnider_2024 (Tabulae Paralytica)",  meta.data = metadata3, min.cells = 10, min.features = 100)

remove_features3 <- nrow(matrix3) - nrow(seurat3)
remove_cells3 <- ncol(matrix3) - ncol(seurat3)

seurat3 <- subset(seurat3, subset = time %in% c("Uninjured", "7dpi"))

remove_features3_f <- nrow(matrix3) - nrow(seurat3)
remove_cells3_f <- ncol(matrix3) - ncol(seurat3)

rm(metadata3, matrix3, barcodes3, features3)
```


Quality control check

Even though methods of the papers indicate that the matrices have already been filtered, we make a small QC check.

```{r}
# Quality control of features
print(paste("Features removed for QC:", remove_features + remove_features2 + remove_features3))
paste("Features:", nrow(seurat) + nrow(seurat2) + nrow(seurat3))

# Quality control of cells
print(paste("Cells removed for QC:", remove_cells + remove_cells2 + remove_cells3))
print(paste("Cells of other experimental conditionns:", (remove_cells_f + remove_cells2_f + remove_cells3_f) - (remove_cells + remove_cells2 + remove_cells3)))
paste("Cells:", ncol(seurat) + ncol(seurat2) + ncol(seurat3))

```


```{r}
save.image(file = "~/R/Data/Seurats_Acute.RData")
```
