==> 0.2 Meta-analysis_Sub-Chronic.Rmd <==
---
title: "Meta-Analysis-SubAcute"
output: html_document
date: "2025-03-07"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

```{r}
library(Seurat)
library(Matrix)
library(biomaRt)
library(AnnotationDbi)
library(org.Mm.eg.db)
library(ggplot2)
library(dplyr)
library(stringr)
library(readxl)
library(openxlsx)
library(patchwork)
library(tibble)
```

1. Dataset 1 - Matson 2022

```{r}

matrix2 <- readRDS("C:/Users/afornies/Downloads/Dataset2.rds")

counts2 <- matrix2@assays[["RNA"]]@counts

metadata2 <- matrix2@meta.data

metadata2$orig.ident <- "Matson_2022 (Single cell Atlas)"

colnames(metadata2)[colnames(metadata2) == "coarse_clusters"] <- "celltype"
colnames(metadata2)[colnames(metadata2) == "condition"] <- "time"
colnames(metadata2)[colnames(metadata2) == "nGene"] <- "nFeature_RNA"
colnames(metadata2)[colnames(metadata2) == "nUMI"] <- "nCount_RNA"

metadata2$time <- gsub("D_3wpi", "21dpi", metadata2$time)
metadata2$time <- gsub("A_Uninj", "Uninjured", metadata2$time)
metadata2$time <- gsub("E_6wpi", "Sub-Chronic", metadata2$time)
metadata2$time <- gsub("B_1dpi", "1dpi", metadata2$time)
metadata2$time <- gsub("C_1wpi", "7dpi", metadata2$time)

colnames(metadata2) <- make.unique(colnames(metadata2))

metadata2 <- metadata2 %>% select(matches("nCount_RNA|nFeature_RNA|time|celltype"))

metadata2 <- metadata2 [, -c(3,4)]

metadata2$technology <- "snRNA"
metadata2$dataset <- "Matson_2022"
```

```{r}
seurat2 <- CreateSeuratObject(counts2, assay = "RNA", project = "Matson_2022 (Single cell Atlas)", meta.data = metadata2, min.cells= 10, min.features = 100)

remove_features2 <- nrow(matrix2) - nrow(seurat2)
remove_cells2 <- ncol(matrix2) - ncol(seurat2)

seurat2 <- subset(seurat2, subset = time %in% c("Uninjured", "Sub-Chronic", "21dpi", "7dpi", "1dpi"))

remove_cells2_f <- ncol(matrix2) - ncol(seurat2)

rm(matrix2, counts2, metadata2)
```

2. Dataset 2 - Squair 2021

```{r}
matrix4 <- readMM("C:/Users/afornies/Downloads/Dataset4.mtx")
barcodes4 <- read.table("C:/Users/afornies/Downloads/barcodes4.txt.gz", header = F)
features4 <- read.delim("C:/Users/afornies/Downloads/features4.txt.gz", header = F)
metadata4 <- read.delim("C:/Users/afornies/Downloads/metadata4.txt")

colnames(matrix4) <- barcodes4$V1
rownames(matrix4) <- features4$V1

rownames(metadata4) <- metadata4$barcode

metadata4 <- metadata4 %>% rename(time = label)
metadata4 <- metadata4 %>% rename(celltype = cell_type)

metadata4$time <- recode(metadata4$time, "6wNT" = "Sub-Chronic", "UN" = "Uninjured")

metadata4 <- metadata4 %>% select(matches("time|celltype"))

metadata4$technology <- "snRNA"
metadata4$dataset <- "Squair_2021"

```

```{r}
seurat4 <- CreateSeuratObject(matrix4, assay = "RNA", project = "Squair_2021",  meta.data = metadata4, min.cells = 10, min.features = 100)

remove_features4 <- nrow(matrix4) - nrow(seurat4)
remove_cells4 <- ncol(matrix4) - ncol(seurat4)

seurat4 <- subset(seurat4, subset = time %in% c("Uninjured", "Sub-Chronic"))

remove_cells4_f <- ncol(matrix4) - ncol(seurat4)

rm(matrix4, barcodes4, metadata4, features4)

```

3. Dataset 3 - Kathe 2022

```{r}

matrix5 <- readMM("C:/Users/afornies/Downloads/Dataset5.mtx")
barcodes5 <- read.table("C:/Users/afornies/Downloads/barcodes5.txt.gz", header = F)
features5 <- read.delim("C:/Users/afornies/Downloads/features5.txt.gz", header = F)
metadata5 <- read.delim("C:/Users/afornies/Downloads/metadata5.txt")


colnames(matrix5) <- barcodes5$V1
rownames(matrix5) <- features5$V1

rownames(metadata5) <- metadata5$barcode

metadata5 <- metadata5 %>% rename(time = label)
metadata5 <- metadata5 %>% rename(celltype = global_cell_type)

metadata5$time <- recode(metadata5$time, "SCI" = "Sub-Chronic")

metadata5 <- metadata5 %>% select(matches("time|celltype"))

metadata5$technology <- "snRNA"
metadata5$dataset <- "Kathe_2022"

```


```{r}
seurat5 <- CreateSeuratObject(matrix5, assay = "RNA", project = "Kathe_2022",  meta.data = metadata5, min.cells = 10, min.features = 100)

remove_features5 <- nrow(matrix5) - nrow(seurat5)
remove_cells5 <- ncol(matrix5) - ncol(seurat5)

seurat5 <- subset(seurat5, subset = time %in% c("Uninjured", "Sub-Chronic"))

remove_cells5_f <- ncol(matrix5) - ncol(seurat5)

rm(matrix5, barcodes5, metadata5, features5)

```

Quality control check

Even though methods of the papers indicate that the matrices have already been filtered, we make a small QC check.

```{r}
# Quality control of features
print(paste("Features removed for QC:", remove_features2 + remove_features4 + remove_features5))
paste("Features:", nrow(seurat2) + nrow(seurat4) + nrow(seurat5))

# Quality control of cells
print(paste("Cells removed for QC:", remove_cells2 + remove_cells4 + remove_cells5))
print(paste("Cells of other experimental conditionns:", (remove_cells2_f + remove_cells4_f + remove_cells5_f) - (remove_cells2 + remove_cells4 + remove_cells5)))
paste("Cells:", ncol(seurat2) + ncol(seurat4) + ncol(seurat5))

```

```{r}
save.image(file = "~/R/Data/Seurats_Chronic.RData")
```
