==> 4. Proteomic Correlation.Rmd <==
---
title: "Other analysis"
output: html_document
date: "2025-04-03"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(readxl)
library(ggplot2)
library(ggvenn)
library(dplyr)
```

## Once I have the DEG list, I can make different analysis to correlate the results with proteomic data.

*ACUTE CONDITION (7 dpi)*

```{r}
transcriptomic <- read_excel("~/R/Downloads/DEG_Acute.xlsx")

interest_genes <- readRDS("~/R/Downloads/Genes_Filtered.rds")

transcriptomic <-  transcriptomic %>% 
  filter(Gene %in% interest_genes)

proteomic <- read_excel("~/R/Data/MAIN_ECM_Proteome.xlsx")

proteomic <- proteomic %>% 
  filter(Gene %in% interest_genes)

# CORRELATION 
transcriptomic <- transcriptomic [, c("Gene", "avg_log2FC","p_val_adj")]

proteomic <- proteomic [, c("Gene", "logFC_Acute_Ctrl", "adj.P.Val_Acute_Ctrl")]

correlation_table <- merge(transcriptomic, proteomic, by = "Gene", all = T)

colnames(correlation_table)[2] <- "log2FC_Transcriptomics"
colnames(correlation_table)[3] <- "p_value_Transcriptomics"
colnames(correlation_table)[4] <- "log2FC_Proteomics"
colnames(correlation_table)[5] <- "p_value_Proteomics"

correlation_table$log2FC_Transcriptomics[is.na(correlation_table$log2FC_Transcriptomics)] <- 0
correlation_table$p_value_Transcriptomics[is.na(correlation_table$p_value_Transcriptomics)] <- 1

correlation_table$log2FC_Proteomics <- as.numeric(correlation_table$log2FC_Proteomics)
correlation_table$p_value_Proteomics <- as.numeric(correlation_table$p_value_Proteomics)

correlation_table1 <- correlation_table %>% 
  mutate(Tendence = case_when(
    log2FC_Transcriptomics > 0.585 & log2FC_Proteomics > 0.585 ~ "Both upregulated",
    log2FC_Transcriptomics < -0.585 & log2FC_Proteomics < -0.585 ~ "Both downregulated",
    log2FC_Transcriptomics > 0.585 & log2FC_Proteomics < -0.585 ~ "Transcriptomics up, Proteomics down",
    log2FC_Transcriptomics < -0.585 & log2FC_Proteomics > 0.585 ~ "Transcriptomics down, Proteomics up",
    TRUE ~ "Neutral" 
  ))

```

*SUB-CHRONIC CONDITION (4-6 wpi)*

```{r}
transcriptomic <- read_excel("~/R/Downloads/DEG_Sub-Chronic.xlsx")

interest_genes <- readRDS("~/R/Downloads/Genes_Filtered.rds")

transcriptomic <-  transcriptomic %>% 
  filter(Gene %in% interest_genes)

proteomic <- read_excel("~/R/Data/MAIN_ECM_Proteome.xlsx")

proteomic <- proteomic %>% 
  filter(Gene %in% interest_genes)

# CORRELATION
transcriptomic <- transcriptomic [, c("Gene", "avg_log2FC","p_val_adj")]

proteomic <- proteomic [, c("Gene", "logFC_Chronic_Ctrl", "adj.P.Val_Chronic_Ctrl")]

correlation_table <- merge(transcriptomic, proteomic, by = "Gene", all = T)

colnames(correlation_table)[2] <- "log2FC_Transcriptomicsc"
colnames(correlation_table)[3] <- "p_value_Transcriptomics"
colnames(correlation_table)[4] <- "log2FC_Proteomics"
colnames(correlation_table)[5] <- "p_value_Proteomics"

correlation_table$log2FC_Transcriptomics[is.na(correlation_table$log2FC_Transcriptomics)] <- 0
correlation_table$p_value_Transcriptomics[is.na(correlation_table$p_value_Transcriptomics)] <- 1

correlation_table$log2FC_Proteomics <- as.numeric(correlation_table$log2FC_Proteomics)
correlation_table$p_value_Proteomics <- as.numeric(correlation_table$p_value_Proteomics)

correlation_table2 <- correlation_table %>% 
  mutate(Tendence = case_when(
    log2FC_Transcriptomics > 0.585 & log2FC_Proteomics > 0.585 ~ "Both upregulated",
    log2FC_Transcriptomics < -0.585 & log2FC_Proteomics < -0.585 ~ "Both downregulated",
    log2FC_Transcriptomics > 0.585 & log2FC_Proteomics < -0.585 ~ "Transcriptomics up, Proteomics down",
    log2FC_Transcriptomics < -0.585 & log2FC_Proteomics > 0.585 ~ "Transcriptomics down, Proteomics up",
    TRUE ~ "Neutral"
  ))


```

```{r}
# Acute
prefix <- "Acute_"

colnames(correlation_table1) <- sapply(colnames(correlation_table1), function(x) ifelse(x != "Gene", paste0(prefix, x), x))

# Sub-Chronic
prefix <- "SubChronic_"

colnames(correlation_table2) <- sapply(colnames(correlation_table2), function(x) ifelse(x != "Gene", paste0(prefix, x), x))

# Merge and Data tidy
correlation_table <- merge(correlation_table1, correlation_table2, by = "Gene")

correlation_table <- correlation_table %>% 
  dplyr::select(matches("Gene|^Acute|^Sub"))

correlation_table <- correlation_table %>% 
  dplyr::select(matches("Gene|Transcriptomics$|Proteomics$"))

delete_cols <- grep("^\\w+_p_value_\\w+$", colnames(correlation_table), value = TRUE)

correlation_table <- correlation_table[, !colnames(correlation_table) %in% delete_cols]

# Prepare for ggpplot
long_data <- correlation_table %>%
  pivot_longer(cols = -Gene, 
               names_to = c("Condition", "Type"), 
               names_pattern = "(.*)_(.*)",
               values_to = "log2FC")
long_data$Condition <- gsub("_log2FC", "", long_data$Condition)

uninjured_data <- expand.grid(
  Gene = unique(long_data$Gene),
  Condition = "Uninjured",
  Type = unique(long_data$Type),
  log2FC = 0
)
long_data <- bind_rows(long_data, uninjured_data) %>%
  mutate(Condition = factor(Condition, levels = c("Uninjured", "Acute", "SubChronic")))

long_data <- long_data %>%
  mutate(Condition = recode(Condition,
                            "Acute" = "7 dpi",
                            "SubChronic" = "4-6 wpi",
                            "Uninjured" = "Before-Injury"))

saveRDS(long_data, file = "~/R/Downloads/Data_Omics.rds")


```

```{r}
acute_markers <- readRDS("~/R/Downloads/Genes_Acute.rds")

long_data_early <- long_data %>%
  filter(Gene %in% acute_markers)

subchronic_markers <- readRDS("~/R/Downloads/Genes_Subchronic.rds")

long_data_late <- long_data %>%
  filter(Gene %in% subchronic_markers)
```

ALL ACUTE AND SUB-CHRONIC MARKERS

```{r}
color_pal <- c("Transcriptomics" = "tomato", "Proteomics" = "steelblue")

# Acute candidate markers
layout_proteomica_early <- ggplot(long_data_early, aes(x = Condition, y = log2FC, color = Type, group = Type)) +
  geom_line() +
  geom_point() + 
  scale_color_manual(values = color_pal) +
  facet_wrap(~ Gene, scales = "free_y") +
  labs(
    title = "Comparison of Transcriptomic and Proteomic Profiles - Acute Biomarkers",
    x = "Condition", 
    y = "log2 Fold Change",
    color = "Data Type"
  ) +
  theme_minimal(base_family = "sans") +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1, size = 10),
    axis.text.y = element_text(size = 10.5),
    legend.position = "right",
    legent.text = element_text(size = 12),
    strip.text = element_text(size = 14),
    plot.title = element_text(face = "bold", size = 16, hjust = 0.5)) +
  geom_hline(yintercept = 0, linetype = "dashed", color = "tomato") 
  
layout_proteomica_early <- layout_proteomica_early 

ggsave("~/R/Figures2/Layout Proteomics-Transcriptomics Acute Markers.png", layout_proteomica_early, dpi = 320, height = 10, width = 14)

# Subchronic candidate markers
layout_proteomica_late <- ggplot(long_data_late, aes(x = Condition, y = log2FC, color = Type, group = Type)) +
  geom_line() +
  geom_point() + 
  scale_color_manual(values = color_pal) +
  facet_wrap(~ Gene, scales = "free_y") +
  labs(
    title = "Comparison of Transcriptomic and Proteomic Profiles - Sub-Chronic Biomarkers",
    x = "Condition", 
    y = "log2 Fold Change",
    color = "Data Type"
  ) +
  theme_minimal(base_family = "sans") +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1, size = 10),
    axis.text.y = element_text(size = 10.5),
    legend.position = "right",
    legent.text = element_text(size = 12),
    strip.text = element_text(size = 14),
    plot.title = element_text(face = "bold", size = 16, hjust = 0.5))  +
  geom_hline(yintercept = 0, linetype = "dashed", color = "tomato")
  
layout_proteomica_late <- layout_proteomica_late

ggsave("~/R/Figures2/Layout Proteomics-Transcriptomics Subchronic Markers.png", layout_proteomica_late, dpi = 320, height = 10, width = 14)

general_layout <- layout_proteomica_early / layout_proteomica_late
general_layout

ggsave("~/R/Figures2/Layout Proteomics-Transcriptomics General Layout.png", general_layout, dpi = 320, height = 20, width = 14)

```

SPECIFIC PER CORRELATION WITH PROTEOMICS

```{r}
# Classification by correlation at different times
wide_data <- long_data %>%
  pivot_wider(names_from = Type, values_from = log2FC) %>%
  rename(
    transcript = Transcriptomics,
    protein = Proteomics
  ) %>%
  mutate(
    sign_match = case_when(
      transcript > 0.585 & protein > 0.585 ~ "pos",
      transcript < -0.585 & protein < -0.585 ~ "neg",
      TRUE ~ "no"
    )
  )

gene_class <- wide_data %>%
  dplyr::select(Gene, Condition, sign_match) %>%
  pivot_wider(names_from = Condition, values_from = sign_match) %>%
  mutate(
    correlation_group = case_when(
      `7 dpi` %in% c("pos", "neg") & `4-6 wpi` %in% c("pos", "neg") ~ "correlation in both",
      `7 dpi` %in% c("pos", "neg") & `4-6 wpi` == "no" ~ "correlation in 7 dpi",
      `7 dpi` == "no" & `4-6 wpi` %in% c("pos", "neg") ~ "correlation in 4-6 wpi",
      TRUE ~ "no correlation"
    )
  ) %>%
  dplyr::select(Gene, correlation_group)

# Add classification
long_data_with_group <- long_data %>%
  left_join(gene_class, by = "Gene")

# No
long_data_no <- long_data_with_group %>%
  filter(correlation_group == "no correlation")

# Acute
long_data_acute <- long_data_with_group %>%
  filter(correlation_group == "correlation in 7 dpi") %>%
  filter(Gene %in% acute_markers)

# Chronic
long_data_chronic <- long_data_with_group %>%
  filter(correlation_group == "correlation in 4-6 wpi")  %>%
  filter(Gene %in% subchronic_markers)

# Both
long_data_both <- long_data_with_group %>%
  filter(correlation_group == "correlation in both")

long_data_acute$Type <- factor(long_data_acute$Type, levels = c("Transcriptomics", "Proteomics"))
long_data_chronic$Type <- factor(long_data_chronic$Type, levels = c("Transcriptomics", "Proteomics"))
long_data_both$Type <- factor(long_data_both$Type, levels = c("Transcriptomics", "Proteomics"))

```

```{r}
# Visualization

# Custom color palette
color_pal <- c("Transcriptomics" = "tomato", "Proteomics" = "steelblue")

# -------------------------------
# ACUTE 
# -------------------------------
long_data_acute$Gene <- factor(long_data_acute$Gene, levels = c("C1qa","C1qb","C1qc","Col3a1","Cxcl12","Hapln2","Lgi1","Spock3"))

layout_proteomica_acute <- ggplot(long_data_acute %>% filter(Condition == "7 dpi"),
                                  aes(x = Gene, y = log2FC, fill = Type)) +
  geom_col(position = "dodge") +
  scale_fill_manual(values = color_pal) +
  labs(
    title = "Transcriptomic and Proteomic Profiles - Acute Dysregulation",
    x = "Gene", y = "log2 Fold Change", fill = "Data Type"
  ) +
  theme_minimal(base_family = "sans") +
  theme(
    axis.text.x = element_text(angle = 0, hjust = 0.5, size = 12),
    axis.title.x = element_blank(),
    axis.text.y = element_text(size = 7),
    plot.title = element_text(hjust = 0.5, size = 16, face = "bold"),
    legend.position = "right",
    legent.text = element_text(size = 12)
  ) +
  geom_hline(yintercept = 0, linetype = "dashed", color = "tomato")

ggsave("~/R/Figures2/Layout Proteomics-Transcriptomics_Acute.png", layout_proteomica_acute, dpi = 320, height = 7, width = 9)

# -------------------------------
# SUB-CHRONIC 
# -------------------------------
layout_proteomica_chronic <- ggplot(long_data_chronic %>% filter(Condition == "4-6 wpi"), aes(x = Gene, y = log2FC, fill = Type)) +
  geom_col(position = "dodge") +
  scale_fill_manual(values = color_pal) +
  labs(
    title = "Transcriptomic and Proteomic Profiles - Chronic Dysregulation",
    x = "Gene", y = "log2 Fold Change", fill = "Data Type"
  ) +
  theme_minimal(base_family = "sans") +
  theme(
    axis.text.x = element_text(angle = 0, hjust = 0.5, size = 16),
    axis.title.x = element_blank(),
    axis.text.y = element_text(size = 13),
    legend.text = element_text(size = 14),        
    legend.title = element_text(size = 16, face = "bold"), 
    strip.text = element_text(size = 14),
    plot.title = element_text(hjust = 0.5, size = 16, face = "bold"),
    legend.position = "right",
    legent.text = element_text(size = 12)
  ) +
  geom_hline(yintercept = 0, linetype = "dashed", color = "tomato")

ggsave("~/R/Figures2/Layout Proteomics-Transcriptomics_Chronic.png", layout_proteomica_chronic, dpi = 1080, height = 7, width = 9)

# -------------------------------
# BOTH 
# -------------------------------
layout_proteomica_both <- ggplot(long_data_both, aes(x = Condition, y = log2FC, color = Type, group = Type)) +
  geom_line() +
  geom_point() +
  facet_wrap(~ Gene, scales = "free_y") +
  scale_color_manual(values = color_pal) +
  labs(
    title = "Transcriptomic and Proteomic Profiles - Acute & Sub-Chronic Dysregulation",
    x = "Condition", y = "log2 Fold Change", color = "Data Type"
  ) +
  theme_minimal(base_family = "sans") +
   theme(
    axis.text.x = element_text(angle = 90, hjust = 0.5, size = 12),
    axis.title.x = element_blank(),
    axis.title.y =  element_text(size = 15),
    axis.text.y = element_text(size = 15),
    legend.text = element_text(size = 14),        
    legend.title = element_text(size = 16, face = "bold"), 
    strip.text = element_text(size = 16),
    plot.title = element_text(hjust = 0.5, size = 16, face = "bold"),
    legend.position = "right"
  ) +
  geom_hline(yintercept = 0, linetype = "dashed", color = "tomato")

ggsave("~/R/Figures2/Layout Proteomics-Transcriptomics_Both.png", layout_proteomica_both, dpi = 1080, height = 9, width = 12)

layout_proteomica_acute
layout_proteomica_chronic
layout_proteomica_both

```


```{r}
long_data_acute$Gene <- as.character(long_data_acute$Gene)

genes_corr_both <- unique(long_data_both$Gene)
saveRDS(genes_corr_both, file = "~/R/Downloads/Genes_Corr_Both.rds")

genes_corr_acute <- unique(long_data_acute$Gene)
saveRDS(genes_corr_acute, file = "~/R/Downloads/Genes_Corr_Acute.rds")

genes_corr_subchronic <- unique(long_data_chronic$Gene)
saveRDS(genes_corr_subchronic, file = "~/R/Downloads/Genes_Corr_Subchronic.rds")

genes_corr <- c(genes_corr_both, genes_corr_acute, genes_corr_subchronic)
saveRDS(genes_corr, file = "~/R/Downloads/Important_Genes.rds")
```
