==> 5. Specific Celltypes.Rmd <==
---
title: "Cell types"
output: html_document
date: "2025-05-22"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
library(Seurat)
library(dplyr)
library(tidyverse)
library(readxl)
library(ComplexHeatmap)
library(openxlsx)
library(circlize)
```

Biomarker characterization

```{r}
biomarkers_acute <- readRDS("~/R/Downloads/Genes_Corr_Acute.rds")

combined_results <- read_excel("~/R/Data/DEG_Celltype_7dpi.xlsx")

combined_results <- combined_results %>%
  filter(Gene %in% biomarkers_acute)

heatmap_data <- combined_results %>%
  dplyr::select(Gene, celltype, avg_log2FC) %>%
  pivot_wider(names_from = celltype, values_from = avg_log2FC) %>%
  column_to_rownames("Gene") %>%
  as.matrix()

heatmap_data[is.na(heatmap_data)] <- 0

colnames(heatmap_data)[colnames(heatmap_data) == "Newly formed oligodendrocytes"] <- "NFOLs"

heatmap_7dpi <- Heatmap(
  t(heatmap_data),
  col = colorRamp2(c(-5, 0, 5), c("steelblue", "white", "tomato")),
  name = "log2FC",
  rect_gp = gpar(col = "grey", lwd = 2),
  row_names_gp = gpar(fontsize = 11),
  row_names_side = "left",
  column_names_gp = gpar(fontsize = 12),
  column_names_rot = 90,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_row_names = TRUE,
  clustering_method_columns = "ward.D2",
  clustering_distance_columns = "euclidean",
  column_title = "Acute",
  column_title_gp = gpar(fontsize = 10)
)

heatmap_7dpi

png("~/R/Figures2/Heatmap_Celltype_7dpi.png",  width = 2500, height = 3500, res = 580)
draw(heatmap_7dpi,
  column_title = "          Expression in different cell types of the acute biomarkers",
  column_title_gp = gpar(fontsize = 16, fontface = "bold"))
dev.off()
```

```{r}
biomarkers_subchronic <- readRDS("~/R/Downloads/Genes_Corr_Subchronic.rds")

combined_results <- read_excel("~/R/Data/DEG_Celltype_Chronic.xlsx")

combined_results <- combined_results %>%
  filter(Gene %in% biomarkers_subchronic)

heatmap_data <- combined_results %>%
  dplyr::select(Gene, celltype, avg_log2FC) %>%
  pivot_wider(names_from = celltype, values_from = avg_log2FC) %>%
  column_to_rownames("Gene") %>%
  as.matrix()

heatmap_data[is.na(heatmap_data)] <- 0

colnames(heatmap_data)[colnames(heatmap_data) == "Newly formed oligodendrocytes"] <- "NFOLs"

heatmap_chronic <- Heatmap(
  t(heatmap_data),
  col = colorRamp2(c(-5, 0, 5), c("steelblue", "white", "tomato")),
  name = "log2FC",
  rect_gp = gpar(col = "grey", lwd = 2),
  row_names_gp = gpar(fontsize = 12),
  row_names_side = "left",
  column_names_gp = gpar(fontsize = 12),
  column_names_rot = 90,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_row_names = TRUE,
  clustering_method_columns = "ward.D2",
  clustering_distance_columns = "euclidean",
  column_title = "Sub-chronic",
  column_title_gp = gpar(fontsize = 10)
)

heatmap_chronic

png("~/R/Figures2/Heatmap_Celltype_Acute+Sub.png",  width = 3500, height = 2500, res = 600)
draw(heatmap_7dpi + heatmap_chronic,
  column_title = "                         Expression in different cell types of biomarkers    ",
  column_title_gp = gpar(fontsize = 12, fontface = "bold", hjust = 0.5)
)
dev.off()
```


```{r}
library(readxl)
library(dplyr)
library(tidyr)
library(tibble)
library(circlize)
library(ComplexHeatmap)

# Load gene list
genes <- readRDS("~/R/Downloads/Genes_Corr_Both.rds")

# ---- 7 dpi ----
results_7dpi <- read_excel("~/R/Data/DEG_Celltype_7dpi.xlsx") %>%
  filter(Gene %in% genes)

mat_7dpi <- results_7dpi %>%
  dplyr::select(Gene, celltype, avg_log2FC) %>%
  pivot_wider(names_from = celltype, values_from = avg_log2FC) %>%
  column_to_rownames("Gene")

mat_7dpi[is.na(mat_7dpi)] <- 0

colnames(mat_7dpi)[colnames(mat_7dpi) == "Newly formed oligodendrocytes"] <- "NFOLs"

hm_7dpi <- Heatmap(
  t(as.matrix(mat_7dpi)),
  col = colorRamp2(c(-5, 0, 5), c("steelblue", "white", "tomato")),
  name = "log2FC",
  rect_gp = gpar(col = "grey", lwd = 2),
  row_names_gp = gpar(fontsize = 7),
  row_names_side = "left",
  column_names_gp = gpar(fontsize = 7),
  column_names_rot = 90,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_row_names = TRUE,
  column_title = "Acute",
  column_title_gp = gpar(fontsize = 10),
  clustering_method_columns = "ward.D2",
  clustering_distance_columns = "euclidean"
)

# ---- Chronic ----
results_chronic <- read_excel("~/R/Data/DEG_Celltype_Chronic.xlsx") %>%
  filter(Gene %in% genes)

mat_chronic <- results_chronic %>%
  dplyr::select(Gene, celltype, avg_log2FC) %>%
  pivot_wider(names_from = celltype, values_from = avg_log2FC) %>%
  column_to_rownames("Gene")

mat_chronic[is.na(mat_chronic)] <- 0

# Align row order with 7 dpi
mat_chronic <- mat_chronic[rownames(mat_7dpi), ]

colnames(mat_chronic)[colnames(mat_chronic) == "Newly formed oligodendrocytes"] <- "NFOLs"

hm_chronic <- Heatmap(
  t(as.matrix(mat_chronic)),
  col = colorRamp2(c(-5, 0, 5), c("steelblue", "white", "tomato")),
  name = "log2FC",
  rect_gp = gpar(col = "grey", lwd = 2),
  row_names_gp = gpar(fontsize = 7),
  column_names_gp = gpar(fontsize = 7),
  column_names_rot = 90,
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_row_names = FALSE,
  column_title = "Sub-chronic",
  column_title_gp = gpar(fontsize = 10),
  clustering_method_columns = "ward.D2",
  clustering_distance_columns = "euclidean"
)

# ---- Final plot ----
png("~/R/Figures2/Heatmap_Celltype_Combined.png", width = 3200, height = 2800, res = 600)
draw(
  hm_7dpi + hm_chronic,
  column_title = "                Expression in different cell types of the biomarkers",
  column_title_gp = gpar(fontsize = 10, fontface = "bold")
)
dev.off()

```

Cell type and Time

```{r}
library(dplyr)
library(ggplot2)
library(ggrepel)

celltype_1dpi <- read_excel("~/R/Data/DEG_Celltype_1dpi.xlsx")
celltype_7dpi <- read_excel("~/R/Data/DEG_Celltype_7dpi.xlsx")
celltype_21dpi <- read_excel("~/R/Data/DEG_Celltype_21dpi.xlsx")
celltype_chronic <- read_excel("~/R/Data/DEG_Celltype_4-6wpi.xlsx")

deg_celltype <- bind_rows(celltype_chronic, celltype_1dpi, celltype_21dpi, celltype_7dpi)

interest_genes <- readRDS(file = "~/R/Downloads/Important_Genes.rds")

deg_celltype <- deg_celltype %>%
  filter(Gene %in% interest_genes)

deg_celltype <- deg_celltype %>% dplyr::select(matches("Gene|avg_log2FC|time|celltype"))

new_rows <- deg_celltype %>% distinct(Gene, celltype) %>% mutate(time = "Uninjured", avg_log2FC = 0)
deg_celltype <- bind_rows(deg_celltype, new_rows) %>% arrange(Gene, celltype, time)
deg_celltype$time <- factor(deg_celltype$time, levels = c("Uninjured","1 dpi", "7 dpi", "21 dpi", "4-6 wpi"))

label_data <- deg_celltype %>%
  group_by(Gene, celltype) %>%
  slice_max(order_by = time, n = 1, with_ties = FALSE) %>%
  ungroup()

layout_celltype <- ggplot(deg_celltype, aes(x = time, y = avg_log2FC, group = Gene, color = Gene)) +
  geom_line() +
  geom_point() +
  facet_wrap(. ~ celltype, scales = "free_y") +
  geom_text_repel(data = label_data, aes(label = Gene), size = 3, nudge_x = 0.3, direction = "y", segment.color = "grey50") +
  theme_minimal() +
  labs(x = "Time Post-Injury", y = "log2 Fold Change", title = "Expression Changes Over Time") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_color_viridis_d()

layout_celltype

ggsave("~/R/Figures2/MainCelltypes_Candidates.png", width = 12, height = 8, dpi = 1080)

```
